Splits contextMenu in seperate files, pre-work for #10301
Auditors: @bsclifton

Test Plan:
- check if hambuger menu is working correctly
