+++ app/renderer/components/main/main.js
@@ -37,7 +37,7 @@ const AddEditBookmarkFolder = require('../bookmarks/addEditBookmarkFolder')
 const LoginRequired = require('./loginRequired')
 const ReleaseNotes = require('./releaseNotes')
 const BookmarksToolbar = require('../bookmarks/bookmarksToolbar')
-const ContextMenu = require('../common/contextMenu')
+const ContextMenu = require('../common/contextMenu/contextMenu')
 const PopupWindow = require('./popupWindow')
 const NoScriptInfo = require('./noScriptInfo')
 const CheckDefaultBrowserDialog = require('./checkDefaultBrowserDialog')
